updates film controller to verify user input

+++ src/main/java/com/skilldistillery/film/controllers/FilmController.java
@@ -1,12 +1,13 @@
 package com.skilldistillery.film.controllers;
 
 import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -20,6 +21,7 @@ import com.skilldistillery.film.entities.Film;
 public class FilmController {
 
 	private int filmId = 0;
+	private Film filmForDisplay = null;
 
 	@Autowired
 	private MVCFilmSiteDAO filmDAO;
@@ -29,12 +31,39 @@ public class FilmController {
 	}
 
 	@RequestMapping(path = "findFilmById.do", method = RequestMethod.GET, params = "id")
-	public ModelAndView findFilmByID(@RequestParam("id") int filmId) {
+	public ModelAndView findFilmByID(@RequestParam("id") String filmId) {
 		ModelAndView mv = new ModelAndView();
-		mv.addObject("film", filmDAO.findFilmById(filmId).displayFilm());
-		mv.setViewName("WEB-INF/result.jsp");
-		return mv;
 
+		Pattern p = Pattern.compile("^[0-9]");
+		Matcher m = p.matcher(filmId);
+
+		if (m.find()) {
+
+			int id = Integer.parseInt(filmId);
+
+			if (id != 0) {
+				Film film = filmDAO.findFilmById(id);
+
+				if (film != null) {
+					filmForDisplay = film;
+					mv.addObject("film", filmDAO.findFilmById(id).displayFilm());
+					mv.setViewName("WEB-INF/result.jsp");
+				}
+
+				else {
+					mv.addObject("film", "Invalid Film, try Again!");
+					mv.setViewName("WEB-INF/filmByID.jsp");
+				}
+			}
+
+		}
+
+		else {
+			mv.addObject("film", "Invalid Film, try Again!");
+			mv.setViewName("WEB-INF/filmByID.jsp");
+		}
+
+		return mv;
 	}
 
 	@RequestMapping(path = "createFilm.do", method = RequestMethod.POST)
@@ -54,41 +83,60 @@ public class FilmController {
 		return mv;
 
 	}
-	
+
 	@RequestMapping(path = "searchFilm.do", method = RequestMethod.GET, params = "keyword")
 	public ModelAndView searchFilm(@RequestParam("keyword") String keyword) {
 		ModelAndView mv = new ModelAndView();
 		List<Film> films = filmDAO.findFilmBySearch(keyword);
-		if (films.isEmpty()){
+		if (films.isEmpty()) {
 			mv.addObject("result", "No matching films found!");
 			mv.setViewName("/WEB-INF/status.jsp");
 			return mv;
 		}
-		
+
 		else {
 			mv.addObject("result", "Here is a list of films we found matching your keyword:<br>");
 			mv.addObject("films", films);
 			mv.setViewName("/WEB-INF/status.jsp");
 			return mv;
 		}
-		
+
 	}
+
 	@RequestMapping(path = "deleteFilm.do", method = RequestMethod.POST, params = "id")
-	private ModelAndView deleteFilm(@RequestParam("id") int filmId) {
+	private ModelAndView deleteFilm(@RequestParam("id") String filmId) {
 		ModelAndView mv = new ModelAndView();
-		Film film = filmDAO.findFilmById(filmId);
-		if (filmDAO.deleteFilm(film)) {
-			mv.addObject("result", "Movie was deleted!");
-			mv.setViewName("/WEB-INF/status.jsp");
-			return mv;
+
+		Pattern p = Pattern.compile("^[0-9]");
+		Matcher m = p.matcher(filmId);
+
+		if (m.find()) {
+
+			int id = Integer.parseInt(filmId);
+
+			Film film = filmDAO.findFilmById(id);
+
+			if (film != null) {
+
+				if (filmDAO.deleteFilm(film)) {
+					mv.addObject("result", "Movie was deleted!");
+					mv.setViewName("/WEB-INF/status.jsp");
+				}
+
+				else {
+					mv.addObject("result", "Movie was not deleted!");
+					mv.setViewName("/WEB-INF/result.jsp");
+				}
+			}
 		}
-		
+
 		else {
-			mv.addObject("result", "Movie was not deleted!");
-			mv.setViewName("/WEB-INF/status.jsp");
-			return mv;
+			mv.addObject("film", filmForDisplay.displayFilm());
+			mv.addObject("result", "Invalid Film!");
+			mv.setViewName("/WEB-INF/result.jsp");
 		}
-		
+
+		return mv;
 	}
 
 	@RequestMapping(path = "editFilmForm.do", method = RequestMethod.GET, params = "id")
